Build Windows packages with Visual Studio 2026
GitHub's windows-latest runner image no longer ships Visual Studio 2022,
so the aws-lc-sys CMake configure step fails with 'could not find any
instance of Visual Studio' when the vs2022 compiler packages request the
Visual Studio 17 2022 generator.

Switch the backend recipe variants to the versioned vs compiler packages
pinned to 2026. Replace the compiler dependencies of the compilers
feature with a dev dependency on the pixi package itself: the package
declares compilers = ["rust", "c"] plus its build and host
dependencies, and the [dev] table exposes them in the development
environments. The rust toolchain pin moves to a workspace build variant.

The conda rust compiler packages set CARGO_BUILD_TARGET, which would
make cargo emit binaries into a target-triple subdirectory. Unset it in
the activation script so the binaries stay in target/pixi/{debug,release}
where the scripts and tests expect them.

Since the [dev] table makes every workspace command resolve the build
backend, pin the backend channels explicitly and apply
PIXI_BUILD_BACKEND_OVERRIDE to the whole integration test suite instead
of only the pixi-build tests. The workspace-built backends match the
build backend API of the pixi binary under test, which released backends
from the channels do not necessarily do.
