+++ pixi-build-backends/recipe/variants.yaml
@@ -1,7 +1,7 @@
 # This is a semi distilled variant config from conda-forge.
 #
 # It adds sysroot (2.17) for linux, macosx_deployment_target for osx
-# and uses vs2022 for windows.
+# and uses vs2026 for windows.
 #
 c_stdlib:
   - if: linux
@@ -23,24 +23,28 @@ c_compiler:
   - if: osx
     then: clang
   - if: win
-    then: vs2022
+    then: vs
 c_compiler_version:
   - if: linux
     then: 14
   - if: osx
     then: 19
+  - if: win
+    then: 2026
 cxx_compiler:
   - if: linux
     then: gxx
   - if: osx
     then: clangxx
   - if: win
-    then: vs2022
+    then: vs
 cxx_compiler_version:
   - if: linux
     then: 14
   - if: osx
     then: 19
+  - if: win
+    then: 2026
 
 python:
   - 3.13
