+++ pixi.toml
@@ -14,33 +14,30 @@ platforms = ["linux-64", "win-64", "osx-64", "osx-arm64", "linux-aarch64"]
 preview = ["pixi-build"]
 requires-pixi = ">=0.67"
 
+[workspace.build-variants]
+rust_compiler_version = [">=1.94,<1.95"]
+
 [workspace.target.linux-64.build-variants]
 c_stdlib_version = ["2.17"]
 
 [workspace.target.osx.build-variants]
 c_compiler_version = ["19"]
 
+[workspace.target.win.build-variants]
+c_compiler = ["vs"]
+c_compiler_version = ["2026"]
+
 #
-# Shared feature for C/C++ compilers and build tools
+# Shared feature that exposes the build dependencies of the pixi package,
+# including compilers, in the environment.
 #
 # Use cmake to build aws-lc-sys, because conda's CFLAGS inject -O2 which
 # overrides the -O0 required by jitterentropy-base.c, causing a build failure.
 [feature.compilers.activation.env]
 AWS_LC_SYS_CMAKE_BUILDER = "1"
 
-[feature.compilers.dependencies]
-c-compiler = ">=1.11.0,<2"
-cmake = ">=4,<5"
-cxx-compiler = ">=1.11.0,<2"
-openssl = ">=3.5.4,<4"
-pkg-config = ">=0.29.2,<0.30"
-
-[feature.compilers.target.linux.dependencies]
-c-compiler = ">=1.6.0"
-clang = ">=22.1.3,<23"
-cxx-compiler = ">=1.6.0"
-make = ">=4.4.1,<5"
-mold = ">=2.40.2,<3"
+[feature.compilers.dev]
+pixi = { path = "." }
 
 #
 # Shared feature for Python
@@ -55,7 +52,6 @@ python = ">=3.14,<3.15"
 env.CARGO_TARGET_DIR = "target/pixi"
 
 [feature.rust.dependencies]
-rust = ">=1.94.0,<1.95"
 rust-src = ">=1.94.0,<2"
 
 [feature.rust.target.unix.activation]
@@ -448,17 +444,30 @@ lefthook = { features = ["lefthook"] }
 name = "pixi"
 
 [package.build]
-backend = { name = "pixi-build-rust", version = "*" }
 source = { path = "crates/pixi" }
+
+[package.build.backend]
+channels = [
+  "https://prefix.dev/pixi-build-backends",
+  "https://prefix.dev/conda-forge",
+]
+name = "pixi-build-rust"
+version = "*"
 
 # Use cmake to build aws-lc-sys, because conda's CFLAGS inject -O2 which
 # overrides the -O0 required by jitterentropy-base.c, causing a build failure.
 [package.build.config]
+compilers = ["rust", "c"]
 env = { AWS_LC_SYS_CMAKE_BUILDER = "1" }
 
-[package.target.unix.build-dependencies]
-clang = "*"
-cmake = ">=3,<4"
+[package.build-dependencies]
+cmake = ">=4,<5"
+pkg-config = ">=0.29.2,<0.30"
+
+[package.host-dependencies]
+openssl = ">=3.5.4,<4"
 
 [package.target.linux.build-dependencies]
-make = ">=4,<5"
+clang = ">=22.1.3,<23"
+make = ">=4.4.1,<5"
+mold = ">=2.40.2,<3"
